chore: Move tools into separate modules
