+++ pyproject.toml
@@ -13,9 +13,6 @@ dependencies = [
     "pillow (>=11.2.1,<12.0.0)"
 ]
 
-[project.scripts]
-nc-mcp-server = "nextcloud_mcp_server.server:run"
-
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 asyncio_default_test_loop_scope = "session"
